fhir: move to test-infra (#134)
* fhir: move to test-infra

* fhir: update readme
